refactor: move custom `*TestCase` classes to `NumberNine\Bundle\Test`
