+++ tests/Functional/Repository/CommentRepositoryTest.php
@@ -14,13 +14,13 @@ declare(strict_types=1);
 namespace NumberNine\Tests\Functional\Repository;
 
 use Doctrine\ORM\Tools\Pagination\Paginator as DoctrinePaginator;
+use NumberNine\Bundle\Test\UserAwareTestCase;
 use NumberNine\Entity\Comment;
 use NumberNine\Model\Content\CommentStatusInterface;
 use NumberNine\Model\Pagination\PaginationParameters;
 use NumberNine\Pagination\Paginator;
 use NumberNine\Repository\CommentRepository;
 use NumberNine\Tests\CreateEntitiesHelperTrait;
-use NumberNine\Tests\UserAwareTestCase;
 
 /**
  * @covers \NumberNine\Repository\CommentRepository
