+++ tests/Unit/Form/Admin/Settings/AdminSettingsGeneralFormTypeTest.php
@@ -13,10 +13,10 @@ declare(strict_types=1);
 
 namespace NumberNine\Tests\Unit\Form\Admin\Settings;
 
+use NumberNine\Bundle\Test\FormTestCase;
 use NumberNine\Entity\Post;
 use NumberNine\Form\Admin\Settings\AdminSettingsGeneralFormType;
 use NumberNine\Model\Content\PublishingStatusInterface;
-use NumberNine\Tests\FormTestCase;
 
 /**
  * @internal
