+++ tests/Unit/Twig/Extension/ColorRuntimeTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace NumberNine\Tests\Unit\Twig\Extension;
 
+use NumberNine\Bundle\Test\DotEnvAwareWebTestCase;
 use NumberNine\Exception\InvalidHexColorValueException;
-use NumberNine\Tests\DotEnvAwareWebTestCase;
 use NumberNine\Twig\Extension\ColorRuntime;
 
 /**
