+++ tests/Unit/Twig/Extension/MediaRuntimeTest.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
 
 namespace NumberNine\Tests\Unit\Twig\Extension;
 
+use NumberNine\Bundle\Test\UserAwareTestCase;
 use NumberNine\Entity\ContentEntityRelationship;
 use NumberNine\Entity\MediaFile;
 use NumberNine\Entity\Post;
 use NumberNine\Media\MediaFileFactory;
-use NumberNine\Tests\UserAwareTestCase;
 use NumberNine\Twig\Extension\MediaRuntime;
 
 /**
