+++ tests/Unit/Twig/Extension/RelationshipRuntimeTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace NumberNine\Tests\Unit\Twig\Extension;
 
+use NumberNine\Bundle\Test\DotEnvAwareWebTestCase;
 use NumberNine\Entity\Post;
-use NumberNine\Tests\DotEnvAwareWebTestCase;
 use NumberNine\Twig\Extension\RelationshipRuntime;
 
 /**
