+++ tests/Unit/Twig/Extension/ThemeRuntimeTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace NumberNine\Tests\Unit\Twig\Extension;
 
+use NumberNine\Bundle\Test\UserAwareTestCase;
 use NumberNine\Configuration\ConfigurationReadWriter;
 use NumberNine\Entity\ContentEntityTerm;
 use NumberNine\Entity\Post;
@@ -20,7 +21,6 @@ use NumberNine\Entity\Term;
 use NumberNine\Model\Content\PublishingStatusInterface;
 use NumberNine\Model\General\Settings;
 use NumberNine\Repository\TaxonomyRepository;
-use NumberNine\Tests\UserAwareTestCase;
 use NumberNine\Twig\Extension\ThemeRuntime;
 
 /**
